Refactor: Extract duplicated downloadGame() into shared utility
- Create utils/downloadGame.js with input validation and return value
- Remove duplicate implementations from Create.jsx and Game.jsx
- Both components now import from the shared utility
- Add JSDoc documentation and error handling for invalid input
